rename VGGType to VType and remove VGGUtils.h
