+++ src/Layer/interface/Core/PaintNode.h
@@ -2,9 +2,8 @@
 // #include "Basic/Renderer.hpp"
 #include "Common/Config.h"
 #include "Core/Node.hpp"
-#include "Core/VGGType.h"
+#include "Core/VType.h"
 #include "Core/Geometry.hpp"
-#include "Core/VGGUtils.h"
 #include "Core/Attrs.h"
 #include "Core/RenderState.h"
 #include "Core/Mask.h"
