+++ src/Layer/src/Core/TextNode.cpp
@@ -3,8 +3,7 @@
 #include "Core/FontManager.h"
 #include "Core/Node.hpp"
 #include "Core/PaintNode.h"
-#include "Core/VGGType.h"
-#include "Core/VGGUtils.h"
+#include "Core/VType.h"
 #include "SkiaBackend/SkFontMgrVGG.h"
 #include "include/core/SkCanvas.h"
 #include "include/core/SkFont.h"
